re-wire to share "precision" resistors
Rewire to make the a-b-wiper connections operate in the expected order
(higher code value => higher diff voltage) and therefore also make the
routing simpler, and allow sharing the 33k resistor at least.

The 150k resistor for the second range digipot is balanced by the 220k
bridge resistor, which is "too much" if you're using PWM mode, but
you're resoldering things if you want to use this PWM mode anyway, so
that's ok.